sera 21.01.23

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: jcardina <jcardina@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/21 15:27:00 by jcardina          #+#    #+#             */
-/*   Updated: 2023/01/21 15:33:06 by jcardina         ###   ########.fr       */
+/*   Updated: 2023/01/21 17:15:14 by jcardina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] != '\0' || s2[i] != '\0') && i < n)
